+++ Sprocket/CMakeLists.txt
@@ -26,6 +26,7 @@ find_path(STB_INCLUDE_DIR NAMES stb_image.h)
 add_library(Sprocket STATIC
             Core/Window.cpp
             Core/timer.cpp
+            Core/log.cpp
 
             Audio/Listener.cpp
             Audio/Music.cpp
@@ -70,7 +71,6 @@ add_library(Sprocket STATIC
             Graphics/Rendering/Scene3DRenderer.cpp
             Graphics/Rendering/SkyboxRenderer.cpp
             
-            Utility/Log.cpp
             Utility/Colour.cpp
             Utility/FileBrowser.cpp
             Utility/InputProxy.cpp
