+++ Sprocket/Core/log.cpp
@@ -1,4 +1,4 @@
-#include "Log.h"
+#include "log.h"
 
 #include <fmt/core.h>
 #include <fmt/color.h>
